Define Token enum and global variables
When it comes to implementing a language, the first thing needed is the
ability to process a text file and recognize what it says. The
traditional way to do this is to use a "lexer" (aka 'scanner') to break
the input up into "tokens". Each token returned by the lexer includes a
token code and potentially some metadata (e.g. the numeric value of a
number). First, we define the possibilities:

[code]

Each token returned by our lexer will either be one of the Token enum
values or it will be an ‘unknown’ character like ‘+’, which is returned
as its ASCII value. If the current token is an identifier, the
IdentifierStr global variable holds the name of the identifier. If the
current token is a numeric literal (like 1.0), NumVal holds its value.
Note that we use global variables for simplicity, this is not the best
choice for a real language implementation :).