Create gettok(); add whitespace skipping
The actual implementation of the lexer is a single function named
gettok. The gettok function is called to return the next token from
standard input. Its definition starts as:

[code]

gettok works by calling the C getchar() function to read characters one
at a time from standard input. It eats them as it recognizes them and
stores the last character read, but not processed, in LastChar. The
first thing that it has to do is ignore whitespace between tokens. This
is accomplished with the loop above.