Recognize identifiers and keywords
The next thing gettok needs to do is recognize identifiers and specific
keywords like “def”. Kaleidoscope does this with this simple loop:

[code]

Note that this code sets the ‘IdentifierStr‘ global whenever it lexes an
identifier. Also, since language keywords are matched by the same loop,
we handle them here inline.