Recognize numbers
Numeric values are similar (see previous commit message):

[code]

This is all pretty straight-forward code for processing input. When
reading a numeric value from input, we use the C strtod function to
convert it to a numeric value that we store in NumVal. Note that this
isn’t doing sufficient error checking: it will incorrectly read
“1.23.45.67” and handle it as if you typed in “1.23”. Feel free to
extend it :).