Handle variable, binary op, call exprs in AST
Right now we only create the AST, so there are no useful accessor
methods on them. It would be very easy to add a virtual method to pretty
print the code, for example. Here are the other expression AST node
definitions that we’ll use in the basic form of the Kaleidoscope
language:

[code]

This is all (intentionally) rather straight-forward: variables capture
the variable name, binary operators capture their opcode (e.g. ‘+’), and
calls capture a function name as well as a list of any argument
expressions. One thing that is nice about our AST is that it captures
the language features without talking about the syntax of the language.
Note that there is no discussion about precedence of binary operators,
lexical structure, etc.