Add prototype and function representations in AST
For our basic language, these are all of the expression nodes we’ll
define. Because it doesn’t have conditional control flow, it isn’t
Turing-complete; we’ll fix that in a later installment. The two things
we need next are a way to talk about the interface to a function, and a
way to talk about functions themselves:

[code]

In Kaleidoscope, functions are typed with just a count of their
arguments. Since all values are double precision floating point, the
type of each argument doesn’t need to be stored anywhere. In a more
aggressive and realistic language, the “ExprAST” class would probably
have a type field.

With this scaffolding, we can now talk about parsing expressions and
function bodies in Kaleidoscope.