Add lexer buffering to store current token
Now that we have an AST to build, we need to define the parser code to
build it. The idea here is that we want to parse something like “x+y”
(which is returned as three tokens by the lexer) into an AST that could
be generated with calls like this:

ExprAST *X = new VariableExprAST("x");
ExprAST *Y = new VariableExprAST("y");
ExprAST *Result = new BinaryExprAST('+', X, Y);

In order to do this, we’ll start by defining some basic helper routines:

[code]

This implements a simple token buffer around the lexer. This allows us
to look one token ahead at what the lexer is returning. Every function
in our parser will assume that CurTok is the current token that needs to
be parsed.