Add error handling helper functions
The Error routines are simple helper routines that our parser will use
to handle errors. The error recovery in our parser will not be the best
and is not particular user-friendly, but it will be enough for our
tutorial. These routines make it easier to handle errors in routines
that have various return types: they always return null.

With these basic helper functions, we can implement the first piece of
our grammar: numeric literals.