Add numeric literal parsing
We start with numeric literals, because they are the simplest to
process. For each production in our grammar, we’ll define a function
which parses that production. For numeric literals, we have:

[code]

This routine is very simple: it expects to be called when the current
token is a tok_number token. It takes the current number value, creates
a NumberExprAST node, advances the lexer to the next token, and finally
returns.

There are some interesting aspects to this. The most important one is
that this routine eats all of the tokens that correspond to the
production and returns the lexer buffer with the next token (which is
not part of the grammar production) ready to go. This is a fairly
standard way to go for recursive descent parsers.