Add parenthesis parsing
For a better example, the parenthesis operator is defined like this:

[code]

This function illustrates a number of interesting things about the
parser:

1) It shows how we use the Error routines. When called, this function
expects that the current token is a ‘(‘ token, but after parsing the
subexpression, it is possible that there is no ‘)’ waiting. For example,
if the user types in “(4 x” instead of “(4)”, the parser should emit an
error. Because errors can occur, the parser needs a way to indicate that
they happened: in our parser, we return null on an error.

2) Another interesting aspect of this function is that it uses recursion
by calling ParseExpression (we will soon see that ParseExpression can
call ParseParenExpr). This is powerful because it allows us to handle
recursive grammars, and keeps each production very simple. Note that
parentheses do not cause construction of AST nodes themselves. While we
could do it this way, the most important role of parentheses are to
guide the parser and provide grouping.  Once the parser constructs the
AST, parentheses are not needed.