Add identifier / function call parsing
The next simple production is for handling variable references and
function calls:

[code]

This routine follows the same style as the other routines. (It expects
to be called if the current token is a tok_identifier token). It also
has recursion and error handling. One interesting aspect of this is that
it uses look-ahead to determine if the current identifier is a stand
alone variable reference or if it is a function call expression. It
handles this by checking to see if the token after the identifier is a
‘(‘ token, constructing either a VariableExprAST or CallExprAST node as
appropriate.