Add primary expression switch
Now that we have all of our simple expression-parsing logic in place, we
can define a helper function to wrap it together into one entry point.
We call this class of expressions “primary” expressions, for reasons
that will become more clear later in the tutorial. In order to parse an
arbitrary primary expression, we need to determine what sort of
expression it is:

[code]

Now that you see the definition of this function, it is more obvious why
we can assume the state of CurTok in the various functions. This uses
look-ahead to determine which sort of expression is being inspected, and
then parses it with a function call.