Add binary operator precedence table
Now that basic expressions are handled, we need to handle binary
expressions. They are a bit more complex.

Binary expressions are significantly harder to parse because they are
often ambiguous. For example, when given the string “x+y*z”, the parser
can choose to parse it as either “(x+y)*z” or “x+(y*z)”. With common
definitions from mathematics, we expect the later parse, because “*”
(multiplication) has higher precedence than “+” (addition).

There are many ways to handle this, but an elegant and efficient way is
to use Operator-Precedence Parsing. This parsing technique uses the
precedence of binary operators to guide recursion. To start with, we
need a table of precedences:

[code]

For the basic form of Kaleidoscope, we will only support 4 binary
operators (this can obviously be extended by you, our brave and intrepid
reader). The GetTokPrecedence function returns the precedence for the
current token, or -1 if the token is not a binary operator. Having a map
makes it easy to add new operators and makes it clear that the algorithm
doesn’t depend on the specific operators involved, but it would be easy
enough to eliminate the map and do the comparisons in the
GetTokPrecedence function. (Or just use a fixed-size array).