Start implementing expression parsing
With the helper above defined, we can now start parsing binary
expressions. The basic idea of operator precedence parsing is to break
down an expression with potentially ambiguous binary operators into
pieces. Consider ,for example, the expression "a+b+(c+d)*e*f+g".
Operator precedence parsing considers this as a stream of primary
expressions separated by binary operators. As such, it will first parse
the leading primary expression "a", then it will see the pairs [+, b]
[+, (c+d)] [*, e] [*, f] and [+, g]. Note that because parentheses are
primary expressions, the binary expression parser doesn’t need to worry
about nested subexpressions like (c+d) at all.

To start, an expression is a primary expression potentially followed by
a sequence of [binop,primaryexpr] pairs:

[code]

ParseBinOpRHS is the function that parses the sequence of pairs for us.
It takes a precedence and a pointer to an expression for the part that
has been parsed so far. Note that "x" is a perfectly valid expression:
As such, "binoprhs" is allowed to be empty, in which case it returns the
expression that is passed into it. In our example above, the code passes
the expression for "a" into ParseBinOpRHS and the current token is "+".