Start implementing ParseBinOpRHS
The precedence value passed into ParseBinOpRHS indicates the minimal
operator precedence that the function is allowed to eat. For example, if
the current pair stream is [+, x] and ParseBinOpRHS is passed in a
precedence of 40, it will not consume any tokens (because the precedence
of ‘+’ is only 20). With this in mind, ParseBinOpRHS starts with:

[code]

This code gets the precedence of the current token and checks to see if
if is too low. Because we defined invalid tokens to have a precedence of
-1, this check implicitly knows that the pair-stream ends when the token
stream runs out of binary operators.