Add eating the binop to ParseBinOpRHS
If this check succeeds, we know that the token is a binary operator and
that it will be included in this expression:

[code]

As such, this code eats (and remembers) the binary operator and then
parses the primary expression that follows. This builds up the whole
pair, the first of which is [+, b] for the running example.