Finish ParseBinOpRHS; TODO: higher precedence op
In our example, the current operator is "+" and the next operator is
"+", we know that they have the same precedence. In this case we’ll
create the AST node for "a+b", and then continue parsing:

[code]

In our example above, this will turn "a+b+" into "(a+b)" and execute the
next iteration of the loop, with "+" as the current token. The code
above will eat, remember, and parse "(c+d)" as the primary expression,
which makes the current pair equal to [+, (c+d)]. It will then evaluate
the ‘if’ conditional above with "*" as the binop to the right of the
primary. In this case, the precedence of "*" is higher than the
precedence of "+" so the if condition will be entered.