Add higher precedence op to ParseBinOpRHS
The critical question left here is “how can the if condition parse the
right hand side in full”? In particular, to build the AST correctly for
our example, it needs to get all of “(c+d)*e*f” as the RHS expression
variable. The code to do this is surprisingly simple (code from the
above two blocks duplicated for context):

[code]

At this point, we know that the binary operator to the RHS of our
primary has higher precedence than the binop we are currently parsing.
As such, we know that any sequence of pairs whose operators are all
higher precedence than “+” should be parsed together and returned as
“RHS”. To do this, we recursively invoke the ParseBinOpRHS function
specifying “TokPrec+1” as the minimum precedence required for it to
continue. In our example above, this will cause it to return the AST
node for “(c+d)*e*f” as RHS, which is then set as the RHS of the ‘+’
expression.

Finally, on the next iteration of the while loop, the “+g” piece is
parsed and added to the AST. With this little bit of code (14
non-trivial lines), we correctly handle fully general binary expression
parsing in a very elegant way. This was a whirlwind tour of this code,
and it is somewhat subtle. I recommend running through it with a few
tough examples to see how it works.