Add prototype parsing
This wraps up handling of expressions. At this point, we can point the
parser at an arbitrary token stream and build an expression from it,
stopping at the first token that is not part of the expression. Next up
we need to handle function definitions, etc.

The next thing missing is handling of function prototypes. In
Kaleidoscope, these are used both for ‘extern’ function declarations as
well as function body definitions. The code to do this is
straight-forward and not very interesting (once you’ve survived
expressions):

[code]