Add top-level dispatch loop
Now that we have all the pieces, let’s build a little driver that will
let us actually execute this code we’ve built!

The driver for this simply invokes all of the parsing pieces with a
top-level dispatch loop. There isn’t much interesting here, so I’ll just
include the top-level loop.

[code]

The most interesting part of this is that we ignore top-level
semicolons. Why is this, you ask? The basic reason is that if you type
“4 + 5” at the command line, the parser doesn’t know whether that is the
end of what you will type or not. For example, on the next line you
could type “def foo...” in which case 4+5 is the end of a top-level
expression. Alternatively you could type “* 6”, which would continue the
expression. Having top-level semicolons allows you to type “4+5;”, and
the parser will know you are done.