Add codegen functions to AST classes
In order to generate LLVM IR, we want some simple setup to get started.
First we define virtual code generation (codegen) methods in each AST
class:

[code]

The Codegen() method says to emit IR for that AST node along with all
the things it depends on, and they all return an LLVM Value object.
“Value” is the class used to represent a “Static Single Assignment (SSA)
register” or “SSA value” in LLVM. The most distinct aspect of SSA values
is that their value is computed as the related instruction executes, and
it does not get a new value until (and if) the instruction re-executes.
In other words, there is no way to “change” an SSA value. For more
information, please read up on Static Single Assignment - the concepts
are really quite natural once you grok them.

Note that instead of adding virtual methods to the ExprAST class
hierarchy, it could also make sense to use a visitor pattern or some
other way to model this. Again, this tutorial won’t dwell on good
software engineering practices: for our purposes, adding a virtual
method is simplest.