Add an error method for Value returning functions
The second thing we want is an “Error” method like we used for the
parser, which will be used to report errors found during code generation
(for example, use of an undeclared parameter):

[code]

The static variables will be used during code generation. TheModule is
the LLVM construct that contains all of the functions and global
variables in a chunk of code. In many ways, it is the top-level
structure that the LLVM IR uses to contain code.

The Builder object is a helper object that makes it easy to generate
LLVM instructions. Instances of the IRBuilder class template keep track
of the current place to insert instructions and has methods to create
new instructions.

The NamedValues map keeps track of which values are defined in the
current scope and what their LLVM representation is. (In other words, it
is a symbol table for the code). In this form of Kaleidoscope, the only
things that can be referenced are function parameters. As such, function
parameters will be in this map when generating code for their function
body.

With these basics in place, we can start talking about how to generate
code for each expression. Note that this assumes that the Builder has
been set up to generate code into something. For now, we’ll assume that
this has already been done, and we’ll just use it to emit code.