Add code generation for number expressions
Generating LLVM code for expression nodes is very straightforward: less
than 45 lines of commented code for all four of our expression nodes.
First we’ll do numeric literals:

[code]

In the LLVM IR, numeric constants are represented with the ConstantFP
class, which holds the numeric value in an APFloat internally (APFloat
has the capability of holding floating point constants of Arbitrary
Precision). This code basically just creates and returns a ConstantFP.
Note that in the LLVM IR that constants are all uniqued together and
shared. For this reason, the API uses the “foo::get(...)” idiom instead
of “new foo(..)” or “foo::Create(..)”.