Add code generation for variable expressions
References to variables are also quite simple using LLVM. In the simple
version of Kaleidoscope, we assume that the variable has already been
emitted somewhere and its value is available. In practice, the only
values that can be in the NamedValues map are function arguments. This
code simply checks to see that the specified name is in the map (if not,
an unknown variable is being referenced) and returns the value for it.
In future chapters, we’ll add support for loop induction variables in
the symbol table, and for local variables.