Add code generation for binary expressions
Binary operators start to get more interesting. The basic idea here is that we
recursively emit code for the left-hand side of the expression, then the
right-hand side, then we compute the result of the binary expression. In this
code, we do a simple switch on the opcode to create the right LLVM instruction.

In the example above, the LLVM builder class is starting to show its value.
IRBuilder knows where to insert the newly created instruction, all you have to
do is specify what instruction to create (e.g. with CreateFAdd), which operands
to use (L and R here) and optionally provide a name for the generated
instruction.

One nice thing about LLVM is that the name is just a hint. For instance, if the
code above emits multiple “addtmp” variables, LLVM will automatically provide
each one with an increasing, unique numeric suffix. Local value names for
instructions are purely optional, but it makes it much easier to read the IR
dumps.

LLVM instructions are constrained by strict rules: for example, the Left and
Right operators of an add instruction must have the same type, and the result
type of the add must match the operand types. Because all values in Kaleidoscope
are doubles, this makes for very simple code for add, sub and mul.

On the other hand, LLVM specifies that the fcmp instruction always returns an
‘i1’ value (a one bit integer). The problem with this is that Kaleidoscope wants
the value to be a 0.0 or 1.0 value. In order to get these semantics, we combine
the fcmp instruction with a uitofp instruction. This instruction converts its
input integer into a floating point value by treating the input as an unsigned
value. In contrast, if we used the sitofp instruction, the Kaleidoscope ‘<’
operator would return 0.0 and -1.0, depending on the input value.