Add code generation for call expressions
[code]

Code generation for function calls is quite straightforward with LLVM.
The code above initially does a function name lookup in the LLVM
Module’s symbol table. Recall that the LLVM Module is the container that
holds all of the functions we are JIT’ing. By giving each function the
same name as what the user specifies, we can use the LLVM symbol table
to resolve function names for us.

Once we have the function to call, we recursively codegen each argument
that is to be passed in, and create an LLVM call instruction. Note that
LLVM uses the native C calling conventions by default, allowing these
calls to also call into standard library functions like “sin” and “cos”,
with no additional effort.

This wraps up our handling of the four basic expressions that we have so
far in Kaleidoscope. Feel free to go in and add some more. For example,
by browsing the LLVM language reference you’ll find several other
interesting instructions that are really easy to plug into our basic
framework.