Start adding codegen for functions and prototypes
Code generation for prototypes and functions must handle a number of
details, which make their code less beautiful than expression code
generation, but allows us to illustrate some important points. First,
lets talk about code generation for prototypes: they are used both for
function bodies and external function declarations. The code starts
with:

[code]

This code packs a lot of power into a few lines. Note first that this
function returns a “Function*” instead of a “Value*”. Because a
“prototype” really talks about the external interface for a function
(not the value computed by an expression), it makes sense for it to
return the LLVM Function it corresponds to when codegen’d.

The call to FunctionType::get creates the FunctionType that should be
used for a given Prototype. Since all function arguments in Kaleidoscope
are of type double, the first line creates a vector of “N” LLVM double
types. It then uses the Functiontype::get method to create a function
type that takes “N” doubles as arguments, returns one double as a
result, and that is not vararg (the false parameter indicates this).
Note that Types in LLVM are uniqued just like Constants are, so you
don’t “new” a type, you “get” it.

The final line above actually creates the function that the prototype
will correspond to. This indicates the type, linkage and name to use, as
well as which module to insert into. “external linkage” means that the
function may be defined outside the current module and/or that it is
callable by functions outside the module. The Name passed in is the name
the user specified: since “TheModule” is specified, this name is
registered in “TheModule“s symbol table, which is used by the function
call code above.