Handle multiple function definition in codegen
The Module symbol table works just like the Function symbol table when
it comes to name conflicts: if a new function is created with a name
that was previously added to the symbol table, the new function will get
implicitly renamed when added to the Module. The code above exploits
this fact to determine if there was a previous definition of this
function.

In Kaleidoscope, I choose to allow redefinitions of functions in two
cases: first, we want to allow ‘extern’ing a function more than once, as
long as the prototypes for the externs match (since all arguments have
the same type, we just have to check that the number of arguments
match). Second, we want to allow ‘extern’ing a function and then
defining a body for it. This is useful when defining mutually recursive
functions.

In order to implement this, the code above first checks to see if there
is a collision on the name of the function. If so, it deletes the
function we just created (by calling eraseFromParent) and then calling
getFunction to get the existing function with the specified name. Note
that many APIs in LLVM have “erase” forms and “remove” forms. The
“remove” form unlinks the object from its parent (e.g. a Function from a
Module) and returns it. The “erase” form unlinks the object and then
deletes it.