Throw errors on redefining already defined func
In order to verify the logic above, we first check to see if the
pre-existing function is “empty”. In this case, empty means that it has
no basic blocks in it, which means it has no body. If it has no body, it
is a forward declaration. Since we don’t allow anything after a full
definition of the function, the code rejects this case. If the previous
reference to a function was an ‘extern’, we simply verify that the
number of arguments for that definition and this one match up. If not,
we emit an error.