Finish codegen for function prototypes
The last bit of code for prototypes loops over all of the arguments in
the function, setting the name of the LLVM Argument objects to match,
and registering the arguments in the NamedValues map for future use by
the VariableExprAST AST node. Once this is set up, it returns the
Function object to the caller. Note that we don’t check for conflicting
argument names here (e.g. “extern foo(a b a)”). Doing so would be very
straight-forward with the mechanics we have already used above.