Start implementing function codegen
Code generation for function definitions starts out simply enough: we
just codegen the prototype (Proto) and verify that it is ok. We then
clear out the NamedValues map to make sure that there isn’t anything in
it from the last function we compiled. Code generation of the prototype
ensures that there is an LLVM Function object that is ready to go for
us.