Add entry BB generation to function codegen
Now we get to the point where the Builder is set up. The first line
creates a new basic block (named “entry”), which is inserted into
TheFunction. The second line then tells the builder that new
instructions should be inserted into the end of the new basic block.
Basic blocks in LLVM are an important part of functions that define the
Control Flow Graph. Since we don’t have any control flow, our functions
will only contain one block at this point. We’ll fix this in Chapter 5
:).