Add return statement generation to func codegen
Once the insertion point is set up, we call the CodeGen() method for the
root expression of the function. If no error happens, this emits code to
compute the expression into the entry block and returns the value that
was computed. Assuming no error, we then create an LLVM ret instruction,
which completes the function. Once the function is built, we call
verifyFunction, which is provided by LLVM. This function does a variety
of consistency checks on the generated code, to determine if our
compiler is doing everything right. Using this is important: it can
catch a lot of bugs. Once the function is finished and validated, we
return it.