Finish function codegen: handle error case
The only piece left here is handling of the error case. For simplicity,
we handle this by merely deleting the function we produced with the
eraseFromParent method. This allows the user to redefine a function that
they incorrectly typed in before: if we didn’t delete it, it would live
in the symbol table, with a body, preventing future redefinition.

This code does have a bug, though. Since the PrototypeAST::Codegen can
return a previously defined forward declaration, our code can actually
delete a forward declaration. There are a number of ways to fix this
bug, see what you can come up with!