Add section for funcs that can be extern'd
One interesting application of this is that we can now extend the
language by writing arbitrary C++ code to implement operations. For
example, if we add:

[code]

Now we can produce simple output to the console by using things like:
“extern putchard(x); putchard(120);”, which prints a lowercase ‘x’ on
the console (120 is the ASCII code for ‘x’). Similar code could be used
to implement file I/O, console input, and many other capabilities in
Kaleidoscope.

This completes the JIT and optimizer chapter of the Kaleidoscope
tutorial. At this point, we can compile a non-Turing-complete
programming language, optimize and JIT compile it in a user-driven way.
Next up we’ll look into extending the language with control flow
constructs, tackling some interesting LLVM IR issues along the way.