Add using directives; fix Makefile
A few using directives were missing and stopping the code from
compiling. Now it builds, also thanks to fixing the Makefile.
Apparantly, the order of appearance of the source files and the link
directives matters for clang++, but make does not know that so by
default it generates the compile invokation command in the wrong order.
Good to know.