Add FunctionPassManager and basic optimizations
LLVM provides many optimization passes, which do many different sorts of
things and have different tradeoffs. Unlike other systems, LLVM doesn’t
hold to the mistaken notion that one set of optimizations is right for
all languages and for all situations. LLVM allows a compiler implementor
to make complete decisions about what optimizations to use, in which
order, and in what situation.

As a concrete example, LLVM supports both “whole module” passes, which
look across as large of body of code as they can (often a whole file,
but if run at link time, this can be a substantial portion of the whole
program). It also supports and includes “per-function” passes which just
operate on a single function at a time, without looking at other
functions. For more information on passes and how they are run, see the
How to Write a Pass document and the List of LLVM Passes.

For Kaleidoscope, we are currently generating functions on the fly, one
at a time, as the user types them in. We aren’t shooting for the
ultimate optimization experience in this setting, but we also want to
catch the easy and quick stuff where possible. As such, we will choose
to run a few per-function optimizations as the user types the function
in. If we wanted to make a “static Kaleidoscope compiler”, we would use
exactly the code we have now, except that we would defer running the
optimizer until the entire file has been parsed.

In order to get per-function optimizations going, we need to set up a
FunctionPassManager to hold and organize the LLVM optimizations that we
want to run. Once we have that, we can add a set of optimizations to
run. The code looks like this:

[code]

This code defines a FunctionPassManager, “OurFPM”. It requires a pointer
to the Module to construct itself. Once it is set up, we use a series of
“add” calls to add a bunch of LLVM passes. The first pass is basically
boilerplate, it adds a pass so that later optimizations know how the
data structures in the program are laid out. The “TheExecutionEngine”
variable is related to the JIT, which we will get to in the next
section.

In this case, we choose to add 4 optimization passes. The passes we
chose here are a pretty standard set of “cleanup” optimizations that are
useful for a wide variety of code. I won’t delve into what they do but,
believe me, they are a good starting place :).