Replace TheModule with JITHelper; extend JIT class
The JIT class has two new functions: getFunction and getModuleForNewFunction.
The first one - getFunction - searches through all Modules that are already
compiled by the JIT helper and if the function is there a prototype for external
linkage is generated and returned. If the function is already defined in the
current module, though, an error message is displayed. If, on the other hand,
the function is just defined in the current module it is just returned. If the
function is not found anywhere a NULL pointer is returned, indicating an error.

The getModuleForNewFunction checks if there is a Module that hasn't been JITed
and returns it if this is the case. Otherwise, it creates a new Module with a
unique name, saves it as the currently JITed Module and returns it.