Replace abstract "TheExecutionEngine" with JIT
The JIT helper class now has a vector of engines and modules and adds
new engines to the list every time one is needed. The engines act as a
storage for already JITed functions.

Also, the JIT helper compiles all queued functions in the current Module
and adds it to the list of JITed modules, once a pointer to any of its
functions is required. This is a lazy compilation scheme.

The difference between the two is that the Modules vector contains LLVM
representations of the functions, whereas the Engines vector contains
compiled versions of them.