Add HelpingMemoryManager class & getSymbolAddress
The HelpingMemoryManager class extends SectionMemoryManager and provides
the getSymbolAddress function which calls the identically named function
on the master MCJITHelper class. That function crawls all compiled
modules and returns the address of the sought symbol. This class is used
when creating new Engines as the designated MemoryManager. It helps with
cross linking symbols across Modules.