Start codegen for if/then/else
Now that we have it parsing and building the AST, the final piece is
adding LLVM code generation support. This is the most interesting part
of the if/then/else example, because this is where it starts to
introduce new concepts. All of the code above has been thoroughly
described in previous chapters.

We need to generate a Phi operation in order to implement conditional
branches. We choose to do it directly as it is really easy to do so.

In order to generate code for this, we implement the Codegen method for
IfExprAST:

[code]

This code is straightforward and similar to what we saw before. We emit
the expression for the condition, then compare that value to zero to get
a truth value as a 1-bit (bool) value.