Create and add BBs for if/then/else
This code creates the basic blocks that are related to the if/then/else
statement, and correspond directly to the blocks in the example above.
The first line gets the current Function object that is being built. It
gets this by asking the builder for the current BasicBlock, and asking
that block for its “parent” (the function it is currently embedded
into).

Once it has that, it creates three blocks. Note that it passes
“TheFunction” into the constructor for the “then” block. This causes the
constructor to automatically insert the new block into the end of the
specified function. The other two blocks are created, but aren’t yet
inserted into the function.

Once the blocks are created, we can emit the conditional branch that
chooses between them. Note that creating new blocks does not implicitly
affect the IRBuilder, so it is still inserting into the block that the
condition went into. Also note that it is creating a branch to the
“then” block and the “else” block, even though the “else” block isn’t
inserted into the function yet. This is all ok: it is the standard way
that LLVM supports forward references.