Emit then value in ThenBB
After the conditional branch is inserted, we move the builder to start
inserting into the “then” block. Strictly speaking, this call moves the
insertion point to be at the end of the specified block. However, since
the “then” block is empty, it also starts out by inserting at the
beginning of the block. :)

Once the insertion point is set, we recursively codegen the “then”
expression from the AST. To finish off the “then” block, we create an
unconditional branch to the merge block. One interesting (and very
important) aspect of the LLVM IR is that it requires all basic blocks to
be “terminated” with a control flow instruction such as return or
branch. This means that all control flow, including fall throughs must
be made explicit in the LLVM IR. If you violate this rule, the verifier
will emit an error.

The final line here is quite subtle, but is very important. The basic
issue is that when we create the Phi node in the merge block, we need to
set up the block/value pairs that indicate how the Phi will work.
Importantly, the Phi node expects to have an entry for each predecessor
of the block in the CFG. Why then, are we getting the current block when
we just set it to ThenBB 5 lines above? The problem is that the “Then”
expression may actually itself change the block that the Builder is
emitting into if, for example, it contains a nested “if/then/else”
expression. Because calling Codegen recursively could arbitrarily change
the notion of the current block, we are required to get an up-to-date
value for code that will set up the Phi node.