Add PHI to MergeBB; emit MergeBB
Now that the ‘then’ and ‘else’ blocks are emitted, we can finish up with
the merge code:

[code]

The first two lines here are now familiar: the first adds the “merge”
block to the Function object (it was previously floating, like the else
block above). The second block changes the insertion point so that newly
created code will go into the “merge” block. Once that is done, we need
to create the PHI node and set up the block/value pairs for the PHI.

Finally, the CodeGen function returns the phi node as the value computed
by the if/then/else expression. In our example above, this returned
value will feed into the code for the top-level function, which will
create the return instruction.

Overall, we now have the ability to execute conditional code in
Kaleidoscope. With this extension, Kaleidoscope is a fairly complete
language that can calculate a wide variety of numeric functions. Next up
we’ll add another useful expression that is familiar from non-functional
languages...